+++ debug/params.go
@@ -6,11 +6,11 @@ import (
 	"go.uber.org/dig"
 )
 
-// ServerParams provides a set of dependencies for a tracer constructor.
+// ServerParams provides a set of dependencies for a debug server constructor.
 type ServerParams struct {
 	dig.In
 
-	Config    *Config    `optional:"true"`
+	Config    *Config
 	Logger    log.Logger `optional:"true"`
 	Lifecycle fxt.Lifecycle
 }