+++ http/config.go
@@ -1,6 +1,6 @@
 package http
 
-// Config holds a list of options used during the debug server construction.
+// Config holds a list of options used during the http server construction.
 type Config struct {
 	Network string
 	Addr    string