+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Do not require a `net.Addr` in debug server config
 - Debug server is started without goph/serverz
+- Make configurations non-optional options
 
 
 ## 0.2.0 - 2017-10-12